use simple int for internal projection state flag
As we're not compacting our user node data into a persistent
structure and trying to minimize storage space, just use a simple int
for our state flag.

+++ lib/projfs.c
@@ -222,8 +222,7 @@ static int remove_xattr_projflag(int fd)
 struct node_userdata
 {
 	int fd;
-
-	uint8_t proj_flag;
+	int proj_flag;
 };
 
 /**
